+++ .github/workflows/ci.yml
@@ -1,0 +1,31 @@
+name: CI
+
+on:
+  push:
+    branches: [ main, master, work ]
+  pull_request:
+
+jobs:
+  build-test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.11'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install -r requirements.txt
+          python -m pip install pytest pytest-qt mypy ruff black
+      - name: Lint
+        run: |
+          ruff src tests
+          black --check src tests
+          mypy src
+      - name: Run tests
+        run: |
+          python -m pytest -q
+      - name: Build
+        run: |
+          python build.py
