+++ requirements.txt
@@ -18,5 +18,3 @@ black==23.11.0
 # Типы для mypy
 types-setuptools==68.2.0.2
 
-# Дополнительные утилиты
-typing-extensions==4.8.0
